ref(slack): Organize unfurlers into modules
